pyproject.toml
[project]
name = "teleservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "aiohttp>=3.11.0",
    "discord.py>=2.5.0",
    "faster-whisper>=1.2.1",
    "greenlet>=3.3.1",
    "groq>=1.0.0",
    "peewee>=3.19.0",
    "peewee-migrations>=0.3.32",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "requests>=2.32.5",
    "yt-dlp>=2026.2.4",
    "zai-sdk>=0.2.2",
    "pynacl>=1.6.2",
    "google-search-results>=2.4.2",
]
